+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changelog
 
+## [2.3.1](https://github.com/nullplatform/tofu-modules/compare/v2.3.0...v2.3.1) (2026-05-11)
+
+
+### Bug Fixes
+
+* **security:** use static var.cluster_name in count to avoid unknown at plan time ([#338](https://github.com/nullplatform/tofu-modules/issues/338)) ([a2675f4](https://github.com/nullplatform/tofu-modules/commit/a2675f4548f8f3395517378825e2524485aea885))
+
+## [2.3.0](https://github.com/nullplatform/tofu-modules/compare/v2.2.1...v2.3.0) (2026-05-08)
+
+
+### Features
+
+* **eks:** expose ami_release_version and use_latest_ami_release_version ([#334](https://github.com/nullplatform/tofu-modules/issues/334)) ([1d88c1e](https://github.com/nullplatform/tofu-modules/commit/1d88c1e2b9197b1e893d2cc394ead7707b2ce3bf))
+* **istio:** expose istiod_replicas to guarantee HA for node drains ([#292](https://github.com/nullplatform/tofu-modules/issues/292)) ([05a081f](https://github.com/nullplatform/tofu-modules/commit/05a081f7e175b42280ffe40aa108a90281482857))
+* **nullplatform/scope_definition:** expose scope_configuration_name_override (backward-compatible) ([#328](https://github.com/nullplatform/tofu-modules/issues/328)) ([8ef0b0e](https://github.com/nullplatform/tofu-modules/commit/8ef0b0ef85a038fbf374d1ec959d7227b78e6c50))
+
+## [2.2.1](https://github.com/nullplatform/tofu-modules/compare/v2.2.0...v2.2.1) (2026-05-07)
+
+
+### Bug Fixes
+
+* **nullplatform/asset/ecr:** correct invalid provider version constraint operator ([#332](https://github.com/nullplatform/tofu-modules/issues/332)) ([8467496](https://github.com/nullplatform/tofu-modules/commit/8467496a52c2f20bbd1f715ce929fac32ff5874c))
+
+## [2.2.0](https://github.com/nullplatform/tofu-modules/compare/v2.1.1...v2.2.0) (2026-05-07)
+
+
+### Features
+
+* **nullplatform/asset/ecr:** add configurable cross-account pull policy ([#330](https://github.com/nullplatform/tofu-modules/issues/330)) ([6f4392f](https://github.com/nullplatform/tofu-modules/commit/6f4392f2761561567f6df5024b81815b110fe5ad))
+
+## [2.1.1](https://github.com/nullplatform/tofu-modules/compare/v2.1.0...v2.1.1) (2026-05-07)
+
+
+### Bug Fixes
+
+* istio subnet annotation ([#327](https://github.com/nullplatform/tofu-modules/issues/327)) ([57c2495](https://github.com/nullplatform/tofu-modules/commit/57c2495fd180057d4484b368d25d34adeb2378da))
+
+## [2.1.0](https://github.com/nullplatform/tofu-modules/compare/v2.0.2...v2.1.0) (2026-05-05)
+
+
+### Features
+
+* update nullplatform provider to &gt;= 0.0.86 across all modules ([#322](https://github.com/nullplatform/tofu-modules/issues/322)) ([6b5e5ce](https://github.com/nullplatform/tofu-modules/commit/6b5e5ce1649721eb7010903d2229004d0225f97f))
+
 ## [2.0.2](https://github.com/nullplatform/tofu-modules/compare/v2.0.1...v2.0.2) (2026-05-04)